gRPC: fix goroutine leak if client stream CloseSend is not called.
Outside of a unidirectional Client->Server stream there's no requirement
that CloseSend must be called (and it doesn't make much sense for a
Server->Client stream).

If CloseSend is not called, the RPC span will not be ended and a
goroutine will be leaked.

We saw a large number of goroutines leaked when we enabled gRPC
instrumentation in our service.

Someone else has also reported this issue here:
https://github.com/open-telemetry/opentelemetry-go-contrib/issues/818