Make ECS detector not error when not on platform (#1428)
* Make ECS detector not error when not on platform

The OpenTelemetry specification states "failure to detect any resource
information MUST NOT be considered an error". Update the ECS detector to
not return an error when not running on the ECS platform.

Additionally, return nil instead of an empty resource to accommodate a
quick merge when this detector is used. A nil value means the other
resource this will be merged with will be returned immediately instead
of working through the actual merge.

* Add changes to changelog

* Match ECS detector tests with expected behavior

* Add PR number to changelog

* Remove unused errNotOnECS