Fix test failure
The new version of testify appears able to determine a byte slice of
zeros is empty and is now failing a test that should have always failed.
The line in question is testing if a root span has a non-empty parent
span, which it never will. It seems like this test was copied from line
175 where a child span was being validated to have a valid parent. This
updates validate the same child span (based on index) is non-empty.
Additionally, it checks the other child span has a valid parent as well.