Add bug/enhancement instrgen issue templates
The Go instrumentation team need a way to tag issues related to
instrgen. Given they don't currently have write access to the repository
and would need it to add the label themselves, this is a solution that
allows for these issues to be created without modifying permissions.

A follow up PR is planed to add a similar template for all other "area" labels.