Add PR policy exemption for instrgen
In order to increase development velocity for this new prototype, the
strict restrictions imposed on other modules that have been released in
this repository are lifted. This exemption is only meant to last as long
as the module remains in a prototype phase. Upon its first tagged
release it is expected that these exemption be removed.