Prepare for migration to new runtime metrics (#5747)
Part of #5655

This is a refactoring to prepare for the implementation of the new
runtime metrics. It:

* Adds support for `OTEL_GO_X_DEPRECATED_RUNTIME_METRICS`, which can be
set to `true` or `false` to enable/disable the existing runtime metrics.
It initially defaults to `true` while the new metrics are under
development.
* Moves the existing runtime metrics implementation to its own internal
package, deprecatedruntime, to clearly separate it from the new metrics
being added, and to make the eventual removal easier.

This does not change any of the metrics generated, or the public API
surface of the runtime metrics package.