Remove pdata dependency and use proto-go instead (#5789)
It's a bit weird that this autoexport package has a dependency on the
collector (well, pdata).
I believe it makes more sense to use
[proto-go](https://github.com/open-telemetry/opentelemetry-proto-go).

pdata also has a rather high release cadence, even though it isn't
modified very frequently. So this change should slightly reduce the
number of renovabot PRs.