feat(exporters/autoexport): add support of OTEL_EXPORTER_OTLP_{TRACES,METRICS,LOGS}_PROTOCOL env var (#5807)
This commit adds support for signal-specific environment variables to configure the OTLP protocol used with the exporter.
As stated in the [specification](https://github.com/open-telemetry/opentelemetry-specification/blob/v1.34.0/specification/protocol/exporter.md#configuration-options), you
can now use the following environment variables to configure signal-specific protocols:

- `OTEL_EXPORTER_OTLP_TRACES_PROTOCOL`
- `OTEL_EXPORTER_OTLP_LOGS_PROTOCOL`
- `OTEL_EXPORTER_OTLP_METRICS_PROTOCOL`

The package will first attempt to load the protocol for a signal from the corresponding environment variable. If it is not set or is empty, it will try to determine the protocol from `OTEL_EXPORTER_OTLP_PROTOCOL`. If this is also not defined or is empty, it will fall back to `http/protobuf`.

Signed-off-by: thomasgouveia <gouveia.thomas@outlook.fr>