exporters/autoexport: add support for signal-specific protocols environment variables (#5816)
This PR adds support for signal-specific environment variables to
configure the OTLP protocol used with the exporter.

The following environment variables have been added to configure
signal-specific protocols:

- `OTEL_EXPORTER_OTLP_TRACES_PROTOCOL`
- `OTEL_EXPORTER_OTLP_LOGS_PROTOCOL`
- `OTEL_EXPORTER_OTLP_METRICS_PROTOCOL`

The package will first attempt to load the protocol for a signal from
the corresponding environment variable. If it is not set or is empty, it
will try to determine the protocol from `OTEL_EXPORTER_OTLP_PROTOCOL`.
If this is also not defined or is empty, it will fall back to
`http/protobuf`.

**Note for reviewers**: As you'll see in the code, I use my own
[env](https://github.com/thomasgouveia/goutils/tree/main/env) package
(MIT License) to facilitate environment variable management. If needed,
I can backport the used functions into the code here and remove the
dependency. Let me know.

Closes
https://github.com/open-telemetry/opentelemetry-go-contrib/issues/5807

---------

Signed-off-by: thomasgouveia <gouveia.thomas@outlook.fr>
Co-authored-by: Robert Pająk <pellared@hotmail.com>
Co-authored-by: Tyler Yahn <MrAlias@users.noreply.github.com>