Protect against multiple WriteHeader calls
In the RespWriterWrapper there was a logic issue. The instance tracks
whether WriteHeader has been called or not, but was still attempting to
call it again even when it had already been called.

This updates the internal implementation to recognize if it's already
been called and make any additional attempts a no-op. This avoids errors
bubbling up for "superfluous response.WriteHeader calls".

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>