otelaws: Add finalize middleware after instead of before (#5975)
Fixes
https://github.com/open-telemetry/opentelemetry-go-contrib/issues/3368.

## Summary
In any presigned URL flow, the AWS SDK skips all `Finalize` middlewares
after the
[PresignHTTPMiddleware](https://github.com/aws/aws-sdk-go-v2/blob/a2747951bd66c0b687a0a483195843b73f6448bb/internal/v4a/presign_middleware.go#L36-L38).
So, if we move context propagation after the `PresignHTTPMiddleware`, we
will not include context propagation headers in presigned URL requests,
which should prevent the reported issue with signatures.

## Context
The existing middleware in this package adds propagation headers (e.g.
`traceparent`) to outgoing AWS requests, presumably to be used in things
like AWS X-Ray. However, when making requests for presigned URLs, the
AWS SDK adds a bunch of middleware of its own to requests. The most
important of these is the
[PresignHTTPMiddleware](https://github.com/aws/aws-sdk-go-v2/blob/a2747951bd66c0b687a0a483195843b73f6448bb/internal/v4a/presign_middleware.go#L36-L38).
This does a lot of things, but for the purposes of this PR/issue, there
are two particularly relevant things it does:

### Incorporating request headers into the presigned URL signature
When it is built, the middleware receives a presigner. Then, when it is
actually run, it receives a `*smithyHTTP.Request`, and it's entirely up
to the presigner if the headers on the `*smithyHTTP.Request` are
incorporated in the calculation of the presigned URL's signature. If we
look in the [internals of the default v4
signer](https://github.com/aws/aws-sdk-go-v2/blob/main/internal/v4a/v4a.go#L305),
it appears to just incorporate all existing headers and use them to sign
the URL. This leads to the problem described in the above issue:

> The URL generated from code with **otelaws middlewares** has two
values `host` and `traceparent`, while the one without **otelaws
middlewares** has only `host`.

As observed by the comments in that issue, disabling the context
propagation solves this issue, which makes sense as that removes the
`traceparent` header from the signature calculation.

This leads us to the second relevant behaviour of the
`PresignHTTPMiddleware`.

### Short circuiting remaining middleware
From the comments of the `PresignHTTPMiddleware`:
```golang
// PresignHTTPRequestMiddleware provides the Finalize middleware for creating a
// presigned URL for an HTTP request.
//
// Will short circuit the middleware stack and not forward onto the next
// Finalize handler.
```
In addition to this, the Deserialize middlewares get cleared, as seen
[here](https://github.com/aws/aws-sdk-go-v2/blob/a2747951bd66c0b687a0a483195843b73f6448bb/service/s3/api_client.go#L904).

So, in the presigned URL flow, the last middleware that will get called
will be the `PresignHTTPRequestMiddleware`. This is actually useful for
us now, as if we just ensure our context propagation middleware comes
last in the `Finalize` step, then we can be sure it will never run in
the presign flow. Simply shifting from `middleware.Before` to
`middleware.After` does just that.

I've also renamed the function and moved it so that its order reflects
the order of middleware execution (i.e. [finalize before
deserialize](https://aws.github.io/aws-sdk-go-v2/docs/middleware/))

---------

Co-authored-by: Robert Pająk <pellared@hotmail.com>