config: separate schema version support into separate directories (#6412)
This makes the migration from one version to another explicit in any
downstream dependencies. This will make dealing with breaking changes
much easier in the Collector as it will allow us to have end users
toggle a feature gate to enable support for new schema versions.

Doing this today is not possible as we cannot import both a v0.2.0 and a
v0.3.0 compatible config package. With the change in this PR, this will
be possible.

v0.2.0 directory was a copy based on
355fbbdb83be04e56837c8f4d583dae8bcc137bb

---------

Signed-off-by: Alex Boten <223565+codeboten@users.noreply.github.com>
Co-authored-by: Damien Mathieu <42@dmathieu.com>