otelgin: remove unneeded context save/restore to avoid data race
There seems to be absolutely no reason to save and restore the request
context in a deferred function. The way it was done before could
(and did) lead to a race condition because when c.Next() is called,
the request context could be changed by another goroutine, and then
the deferred function would modify the request while another
goroutine is using it. I can't for the life of me figure out what
benefit the previous code was trying to achieve, other than maybe
trying to minimize the context size in memory, but that seems
unnecessary and not worth the risk of a data race. Without this
change, we cannot run our tests with the race detector enabled.