otelconf: experimental x package
This is an attempt at producing a separate `otelconf/x` package to support
experimental features in the configuration schema. This is doing the bare minimum
of using a `sed` patch to hide all experimental structs from the generated model
code before moving it to the `otelconf` package. This has the upside of being
pretty straight forward to do, but ends up with a lot of duplicate code.

Part of #8203

Signed-off-by: alex boten <223565+codeboten@users.noreply.github.com>